Remove broken helm options
Helm have changed the command syntax for installing and deleting charts with 2.16 and 3.x. This means our install docs no longer work.

We don't need --name on the install since we're doing an `upgrade --install`, which always takes release name first, then chart. The delete command is more complicated, and needs `--purge` when using 2.x, and needs to not have it when using 3.x.

Fixes #87