Run as non-root to work on Openshift
Openshift requires that pods run as non-root - root pods won't start. There's some directories in the container's filesystem that need to be writable to Portieris, so added a little macro to the build that allows it to create the directories it needs with the permissions it needs to run as non-root.

Tested on an IBM Cloud Openshift cluster running Openshift 3.11.

Fixes https://github.com/IBM/portieris/issues/80