Adding creating secret example (#154)
* Revert "--purge has been removed from helm 3"

* Adding creating secret example 

In setting this up using the latest feature of look-aside signatures, last week, I spent a few hours trying to understand how the secret should have been created - particularly the name of the key it was looking for. Naming it anything other than 'key' would not work. By default, i believe it picks up the basename of the file. Either way, adding an example for clarity.

Lastly, when I was trying to figure out how to configure for look-aside, I could not find anything relevant to the statement mentioning 'KeyType' and removed it. The literal word 'KeyType' does not show anywhere in the doc and it was a little confusing to read.

* Update README.md

* Update POLICIES.md

Co-authored-by: Stuart Hayton <sjhx@users.noreply.github.com>