Use a namespace selector for admission webhook (#145)
* Use a namespace selector for admission webhook

This prevents the webhook from being called for the Portieris install namespace, which means that Portieris can recover itself in the case of cluster failure.

Without this an approval from the Portieris webhook is needed to okay scaling itself up. This means that with no pods available, the webhook can't approve the recovery of any Portieris pods, and so the cluster deadlocks.

This change gives the Portieris chart ownership of the portieris install namespace, and labels it in such a way that we can filter for it in the webhook config.

It's configured as an opt out, rather than Istio's which is an opt in. All namespaces without the label are fair game.

By adding the namespace into the chart, it'll be deleted by Helm when the chart gets removed. And adding the label selector means that the label could be added to other namespaces to bypass Portieris. Both of these potential issues have been documented in the readme.

https://github.com/IBM/portieris/issues/112

* Add an install option to skip annotated namespaces

And docs to help people decide whether they want it.

* Switch admissionskip to off by default