Fix issue with logging for the digest
At the point the image Reference is created using the image name, the
digest is not available. So a key is created with image:tag in the map.
When the digest has been obtained from the registry, the key in the map
is mutated from image:tag to image:digest

The logging will then show Decision for image:tag or image:digest dependent
upon where in the admitPod lifecycle a the decision is made.