Add additional validation to owner check.
After this we only return `ErrObjectHasParents` in the case the parent is of a supported kind.

Currently if a resource has an owner of any kind we do not admit, as we expect the admission to happen at the owner level.

In the case where the owner is not a supported kind e.g. a CRD this leads to the resource being created with no admission occuring anywhere in the chain.