Fix test failure
The old nginx image caused test failures because it's a v1 manifest and Docker Hub refuses to let you download them now.

I set the image version to the latest as of today. We only need the container to start once it is approved. This is testing for an unsigned image - I'm betting that if nginx ever do decide to sign their images using simple signing in future, they won't go back and sign old images.

Signed-off-by: Michael Hough <michaelh@uk.ibm.com>